pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.45.2",
    "asyncpg>=0.30.0",
    "chainlit>=2.2.1",
    "click>=8.0.1",
    "colorama>=0.4.6",
    "docker>=7.0.0",
    "ijson>=3.2.3",
    "injector>=0.21.0",
    "ipykernel==6.26.0",
    "jsonschema==4.20.0",
    "matplotlib>=3.4",
    "mcp>=1.2.1",
    "numpy>=1.24.2",
    "openai>=1.42.0",
    "pandas>=2.0.0",
    "plotly>=5.14.1",
    "pre-commit>=2.19.0",
    "pydantic>=2.8.2",
    "pytest>=7.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "scikit-learn>=1.2.2",
    "seaborn>=0.11",
    "tenacity>=8.2.2",
    "urllib3>=1.26.17",
    "vcrpy>=5.0.0",
]
